requirements.txt
tensorflow>=2.10.0
tensorflow-hub>=0.12.0
opencv-python>=4.6.0
numpy>=1.23.0
scikit-image>=0.19.0
tqdm>=4.64.0
pandas>=1.5.0
matplotlib>=3.6.0
rich>=12.0.0
pyyaml>=6.0
scipy>=1.9.0
streamlit>=1.20.0
plotly>=5.13.0
requests
